requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
sqlparse==0.5.3
python-dotenv==1.0.1
whitenoise==6.6.0
reportlab==4.0.7
openpyxl==3.1.2
Pillow==10.1.0
requests==2.31.0
psycopg[binary]==3.2.3
dj-database-url==2.2.0
boto3==1.34.162
django-storages==1.14.4
